al2023-ec2/lgtm-docker/requirements.txt
flask 
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-flask 
opentelemetry-exporter-otlp-proto-grpc 
opentelemetry-exporter-otlp 
opentelemetry-instrumentation-logging 
opentelemetry-instrumentation-requests
